Server: 부동 소수점 오차를 고려해 체력, 대미지, 사거리, 이동 속도를 정수 기반으로 변경
기존에는 float를 사용했는데, byte, ushort, uint를 사용해 정수 기반으로 변경하였음.

+++ Server/MiniServerProject/Domain/Battle/ServerBattleMapSimualtor.cs
@@ -77,7 +77,7 @@ public class ServerBattleMapSimulator : IBattleMapEventHandler
         
     }
 
-    public void OnEntityGetDamage(uint entityId, float damage)
+    public void OnEntityGetDamage(uint entityId, uint damage)
     {
         
     }
